outsourced pickle NN object storing from the database into files

+++ examples/delete_low_score.py
@@ -15,11 +15,12 @@ session = DBSession()
 low_scores = session.query(NNTraining).filter(NNTraining.score < 80.0).all()
 
 for low_score in low_scores:
-    print('Removing %s with score %.2f%%' % (low_score.nn_saved_name, low_score.score))
-    file = os.path.join(NN_MODEL_DIR, low_score.nn_saved_name)
-    session.delete(low_score)
-    try:
-        os.remove(file)
-    except FileNotFoundError:
-        print('Did not find on disk: %s' % file)
+    if low_score.nn_saved_name != 'deleted':
+        print('Removing %s with score %.2f%%' % (low_score.nn_saved_name, low_score.score))
+        file = os.path.join(NN_MODEL_DIR, low_score.nn_saved_name)
+        low_score.nn_saved_name = 'deleted'
+        try:
+            os.remove(file)
+        except FileNotFoundError:
+            print('Did not find on disk: %s' % file)
 session.commit()
